Move MemoryStore to memory.rs
