+++ Cargo.toml
@@ -32,6 +32,7 @@ libc = "0.2.155"
 log = "0.4.21"
 pretty_assertions = "1.4.0"
 ratatui = { version = "0.26.3", features = ["serde", "macros"] }
+rusqlite = { version = "0.32.1", features = ["bundled"] }
 serde = { version = "1.0.203", features = ["derive"] }
 serde_json = "1.0.117"
 serde_with = { version = "3.8.1", features = ["chrono_0_4"] }
